Create README.md
Our team has created a braille sensor that converts digital text as shown on a phone/tablet and other small devices into braille for the visually impaired. Current technology for translating braille generally relies on a braille embosser that create a physical copy in braille while other electronic braille converters are extremely expensive. Our goal throughout this project was to create the most cost efficient, user-friendly, portable product that any individual can use on the go. This model eliminates this middle man and press the text directly onto the users fingertips so they can read text messages, navigational directions, and even webpages.  Our device connects to any electronic device via Bluetooth and remains connected until turned off by the user. This design for our product thus frees user’s auditory capabilities for other needs in their day to day life in addition to fitting in the palm of their hand. We believe that this is just the beginning for this product as there are so many use cases including education, and navigation.